Allow dtype option

+++ DLC_for_WBFM/utils/video_and_data_conversion/import_video_as_array.py
@@ -42,11 +42,11 @@ def get_video_from_ome_file_subset(video_fname,
     return dat
 
 
-def get_single_volume(fname, which_vol, num_slices, alpha=1.0):
+def get_single_volume(fname, which_vol, num_slices, alpha=1.0, dtype='uint8'):
     # Convert to page coordinates
     start_ind = num_slices*which_vol
     key = range(start_ind, start_ind+num_slices)
-    return (alpha*tifffile.imread(fname, key=key)).astype('uint8')
+    return (alpha*tifffile.imread(fname, key=key)).astype(dtype)
 
 
 def get_adjacent_volumes(fname, first_vol, num_slices):
